requirements.txt
altgraph==0.17.4
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
et-xmlfile==1.1.0
h11==0.14.0
idna==3.6
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pefile==2023.2.7
pycparser==2.21
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.31.0
selenium==4.18.1
serpapi==0.1.5
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.1.0
wsproto==1.2.0
